% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_workflow_fun.R
\name{all_in_one}
\alias{all_in_one}
\title{All-in-one function}
\usage{
all_in_one(input, output = NULL, saveCheckpoint = TRUE,
  continueCheckpoint = NULL, flankingRegions = NULL, nrThreads = 2,
  control = all_in_one.control())
}
\arguments{
\item{input}{a path to a \code{.fastq}-file, or directory containing multiple files (NOT IMPLEMENTED) all of which should be analysed.}

\item{output}{a directory where output-files are stored.}
}
\value{
The function returns all relevant objects created which might be useful for further analysis.
}
\description{
The function takes an input file and performs the entire analysis workflow described in (ADD REF).
The function creates a series of objects needed for further analyses.
An output folder can be provided to store the objects as \code{.RData}-files.
}
