% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{baseErrorProbability}
\alias{baseErrorProbability}
\title{The weighted average probability of error for each base.}
\usage{
baseErrorProbability(averageQuality, coverageString, sumCoverage,
  windowSize = 5)
}
\arguments{
\item{averageQuality}{A string containing the phred encoded quality.}

\item{coverageString}{The coverage of the string assumed to be correct.}

\item{sumCoverage}{The sum of the coverage on the marker.}

\item{windowSize}{The size of the window used for weighting the probabilities.}
}
\value{
A vector containing the weighted probability of error for each base.
}
\description{
The weighted average probability of error for each base.
}
