% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR_Region_Identification.R
\docType{class}
\name{extractedReadsList-class}
\alias{extractedReadsList-class}
\title{Extract STR region information}
\description{
Identifies the marker of the read using flanking regions and trims the read to include what is between the flanking regions.
}
