% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR_Region_Identification.R
\docType{class}
\name{extractedReadsListCombined-class}
\alias{extractedReadsListCombined-class}
\title{Combined extract STR region information.}
\description{
Identifies the marker of the read for both the provided and reverse complement flanking regions. The resulting lists are then combined into a single list.
}
