% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR_Region_Identification.R
\docType{class}
\name{extractedReadsListNonCombined-class}
\alias{extractedReadsListNonCombined-class}
\title{Combined extract STR region information.}
\description{
Identifies the marker of the read for both the provided and reverse complement flanking regions.
}
