% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR_Region_Identification.R
\docType{class}
\name{extractedReadsListReverseComplement-class}
\alias{extractedReadsListReverseComplement-class}
\title{Extract STR region information of the reverse complement DNA strand.}
\description{
Identifies the marker of the read using reverse complement flanking regions and trims the read to include what is between the flanking regions.
}
