% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stutter_functions.R
\name{findLeftShoulder}
\alias{findLeftShoulder}
\title{Find left shoulder}
\usage{
findLeftShoulder(stringCoverageGenotypeListObject, motifLength, searchDirection,
  gapOpeningPenalty = 6, gapExtensionPenalty = 1)
}
\description{
Find left shoulder
}

