% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stutter_functions.R
\name{findNeighbours,stringCoverageGenotypeList-method}
\alias{findNeighbours,stringCoverageGenotypeList-method}
\title{Find neighbours}
\usage{
\S4method{findNeighbours}{stringCoverageGenotypeList}(
  stringCoverageGenotypeListObject,
  searchDirection = -1,
  trace = FALSE
)
}
\arguments{
\item{stringCoverageGenotypeListObject}{A \link{stringCoverageGenotypeList-class} object.}

\item{searchDirection}{The direction to search for neighbouring strings. Default is -1, indicating a search for '-1' stutters.}

\item{trace}{Should a trace be shown?}
}
\value{
A 'neighbourList' with the neighbouring strings, in the specified direction, for the identified allele regions.
}
\description{
Generic function for finding neighbouring strings, given identified alleles.
}
