% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stutter_functions.R
\docType{methods}
\name{findStutter,stringCoverageGenotypeList-method}
\alias{findStutter,stringCoverageGenotypeList-method}
\title{Find stutters}
\usage{

  \S4method{findStutter}{stringCoverageGenotypeList}(stringCoverageGenotypeListObject,
  trace = FALSE)
}
\arguments{
\item{stringCoverageGenotypeListObject}{A \link{stringCoverageGenotypeList-class} object.}

\item{trace}{Should a trace be shown?}
}
\value{
A 'neighbourList' with the stutter strings for the identified allele regions.
}
\description{
Given identified alleles it search for '-1' stutters of the alleles.
}
\examples{
library("dplyr")

# The object returned by merging a stringCoverageList-Object
# and a genotypeList-Object.
data("stringCoverageGenotypeList")

stutterList <- findStutter(stringCoverageGenotypeList)
stutterTibble <- do.call("rbind", stutterList) \%>\%
    filter(!is.na(Genotype))

stutterTibble$BlockLengthMissingMotif
stutterTibble$NeighbourRatio
}
