% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\docType{class}
\name{genotypeIdentifiedList-class}
\alias{genotypeIdentifiedList-class}
\title{Genotype list}
\description{
A reduced stringCoverageList restricted to the identified genotypes.
}
