% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{genotypeList}
\alias{genotypeList}
\title{Genotype list}
\format{
A list of \link[tibble]{tibble}'s one for each of the 10 markers, showing which strings are the potential alleles based on the 'Coverage'.
}
\usage{
data("genotypeList")
}
\description{
The identified genotypes of the \link{stringCoverageList} data, created by the  \link{getGenotype} function.
}
\author{
Søren B. Vilsen \email{svilsen@math.aau.dk}
}
\keyword{data}
