% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAlignmentNeighbourScore}
\alias{getAlignmentNeighbourScore}
\title{getAlignmentNeighbourScore}
\usage{
getAlignmentNeighbourScore(parent, neighbours, mismatch, gapOpening,
  gapExtension)
}
\arguments{
\item{parent}{the parent string with which the neighbours are matched.}

\item{neighbours}{potential variations of the parent string.}

\item{gapOpening}{the penalty for opening a new gap (positive value).}

\item{gapExtension}{the penalty for extenting a gap (positive value).}
}
\description{
Gets the alignment score of specified neighbours.
}

