% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygote_balance.R
\name{heterozygoteBalance}
\alias{heterozygoteBalance}
\title{Finds the heterozygote balance.}
\usage{
heterozygoteBalance(stringCoverageGenotypeListObject, trueAllelesObject,
  beliefCol = "Coverage")
}
\arguments{
\item{trueAllelesObject}{a trueAlleles-object created using the \link{trueAlleles}-function.}

\item{beliefCol}{the coloumn in which we believe.}

\item{genotypeStringCoverageListObject}{either a stringCoverageList-object or a stringCoverageGenotypeList-object.}
}
\value{
The function returns an object of class 'heterozygoteBalanceList' or 'heterozygoteBalanceUnknownList'.
}
\description{
Given either a stringCoverageListObject and a list of the class 'trueAllelesList' or 'stringCoverageGenotypeListObject', the function finds if possible the heterozygote balance for every provided marker (the elements of the first list).
}
