% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{identifiedSTRs}
\alias{identifiedSTRs}
\title{Identified STR regions}
\format{
A list with an element for each of the 10 identified markers indicating which sequences were identified for each marker.
}
\usage{
data("identifiedSTRs")
}
\description{
The identified STR regions of the sampleSequences.fastq file, created by the  \link{identifySTRRegions} function.
}
\author{
Søren B. Vilsen \email{svilsen@math.aau.dk}
}
\keyword{data}
