% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R,
%   R/auxiliary_functions.R
\docType{data}
\name{identifyNoise}
\alias{identifyNoise}
\alias{identifyNoise}
\title{Idenfities the noise.}
\format{A list of \link{tibble}'s one for each of the 10 markers, showing which strings can be safely classified as noise based on the 'Coverage'.}
\usage{
identifyNoise(stringCoverageListObject, colBelief = "Coverage",
  thresholdSignal = 0.01)

data("identifyNoise")
}
\arguments{
\item{stringCoverageListObject}{an stringCoverageList-object, created using the \link{stringCoverage}-function.}

\item{colBelief}{the name of the coloumn used for identification.}

\item{thresholdSignal}{threshold applied to the signal (generally the coverage) of every string.}
}
\value{
Returns a list, with an element for every marker in stringCoverageList-object, each element contains the genotype for a given marker.
}
\description{
\code{identifyNoise} takes an stringCoverageList-object and identifies the noise based on a signal threshold for every marker in the provided list.

The identified noise of the \link{stringCoverageList} data, created by the  \link{identifyNoise} function.
}
\examples{
# Strings aggregated by 'stringCoverage()'
data("stringCoverageList")

identifyNoise(stringCoverageList, thresholdSignal = 0.03)
}
\author{
Søren B. Vilsen \email{svilsen@math.aau.dk}
}
\keyword{data}
