% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb_noise_threshold.R
\name{kinb.control}
\alias{kinb.control}
\title{K-Inflated Negative Binomial Control Function}
\usage{
kinb.control(method = "BFGS", maxit = 10000, trace = FALSE, EM = FALSE,
  start = NULL, return.pars.only = FALSE, ...)
}
\value{
Returns a list of default objects for the \link{kinb}-function.
}
\description{
K-Inflated Negative Binomial Control Function
}

