% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kinb_noise_threshold.R
\name{kinb}
\alias{kinb}
\title{K-Inflated Negative Binomial Models}
\usage{
kinb(formula, data, subset, na.action, weights, offset,
  control = kinb.control(), model = TRUE, y = TRUE, x = FALSE,
  truncated = TRUE, kInflation = 1L, ...)
}
\arguments{
\item{control}{a \link{kinb.control} object.}

\item{truncated}{a TRUE/FALSE statement indicating whether the model is (k-1)-truncated.}

\item{kInflation}{the point of inflation. The variable defaults to 1.}
}
\value{
A list of parameter estimates (can return a shorter list if needed).
}
\description{
A function fitting k-inflated negative binomial model, (k-1)-truncated and non-truncated.
}

