% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{LUS}
\alias{LUS}
\title{Longest Uninterrupted Stretch}
\usage{
LUS(s, motifLength = 4, returnType = "numeric")
}
\arguments{
\item{s}{a string of either class: 'character' or 'DNAString'.}

\item{motifLength}{the known motif length of the STR region.}

\item{returnType}{the type of return wanted. Takes three values 'numeric', 'string', or 'fullList' (or any other combination cased letters).}
}
\value{
Depending on returnType it return an object of class 'numeric', 'string', or 'fullLUSList'.
}
\description{
Given a motif length and a string it finds the longest uninterrupted stretch (LUS) of the string.
}
\details{
If returnType is 'numeric', the function returns the numeric value of the LUS.
If returnType is instead chosen as 'string', the function returns "[LUSMotif]LUSValue" i.e. the motif of the LUS and its value.
Lastly if the returnType is set to fullList, the function returns a list of data.frames containing every possible repeat structure their start and the numeric value of the repeat unit length.
}
