% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stutter_functions.R
\docType{class}
\name{neighbourList-class}
\alias{neighbourList-class}
\title{A neighbour list}
\description{
A list of the identified neightbours of the called alleles in a stringCoverageGenotypeList
}
