% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\docType{class}
\name{noiseIdentifiedList-class}
\alias{noiseIdentifiedList-class}
\title{Noise list}
\description{
Creates a flag to the sequences in a stringCoverageList which cloud be classified as noise.
}
