% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{noiseList}
\alias{noiseList}
\title{Noise list}
\format{
A list of \link[tibble]{tibble}'s one for each of the 10 markers, showing which strings can be safely classified as noise based on the 'Coverage'.
}
\usage{
data("noiseList")
}
\description{
The identified noise of the \link{stringCoverageList} data, created by the  \link{identifyNoise} function.
}
\author{
Søren B. Vilsen \email{svilsen@math.aau.dk}
}
\keyword{data}
