% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{phredQualityProbability}
\alias{phredQualityProbability}
\title{Phred to Probability}
\usage{
phredQualityProbability(q)
}
\arguments{
\item{q}{Phred Quality Score.}
}
\value{
Probability of error.
}
\description{
Converts Phred Quality Score to probability.
}

