% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{phredQualityScore}
\alias{phredQualityScore}
\title{Phred Quality Score}
\usage{
phredQualityScore(p)
}
\arguments{
\item{p}{probability of error.}
}
\value{
Phred quality score.
}
\description{
Calculates phred quality score given probability.
}

