% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotSequence.ggplot}
\alias{plotSequence.ggplot}
\title{Plots sequences of a given marker.}
\usage{
plotSequence.ggplot(stringCoverageListObject, marker = NULL,
  control = plotSequence.control())
}
\description{
Plots sequences of a given marker.
}

