% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{phredQualityScore}
\alias{phredQualityScore}
\alias{solexaQualityScore}
\title{Convert probability to quality score}
\usage{
phredQualityScore(p)

solexaQualityScore(p)
}
\arguments{
\item{p}{Probability of error.}
}
\value{
\code{phredQualityScore(p)} returns a Phred quality score.

\code{solexaQualityScore(p)} returns a Solexa quality score.
}
\description{
Calculates the quality score (Phred or Solexa) given a probability of error.
}
\examples{
p <- 1e-3
phredQualityScore(p)
solexaQualityScore(p)
}
