% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{solexaQualityProbability}
\alias{solexaQualityProbability}
\title{Solexa to Probability}
\usage{
solexaQualityProbability(q)
}
\arguments{
\item{q}{Solexa Quality Score.}
}
\value{
Probability of error.
}
\description{
Converts Solexa Quality Score to probability.
}
