% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{solexaQualityScore}
\alias{solexaQualityScore}
\title{Solexa Quality Score}
\usage{
solexaQualityScore(p)
}
\arguments{
\item{p}{probability of error.}
}
\value{
Solexa Quality Score.
}
\description{
Calculates Solexa Quality Score given probability.
}

