% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\name{stringCoverage}
\alias{stringCoverage}
\title{Get string coverage STR identified objects.}
\usage{
stringCoverage(extractedReadsListObject, control = stringCoverage.control())
}
\arguments{
\item{extractedReadsListObject}{an extractedReadsList-object, created using the \link{identifySTRRegions}-function.}

\item{control}{an \link{stringCoverage.control}-object.}
}
\value{
Returns a list, with an element for every marker in extractedReadsList-object, each element contains the string coverage of all unique strings of a given marker.
}
\description{
\code{stringCoverage} takes an extractedReadsList-object and finds the coverage of every unique string for every marker in the provided list.
}

