% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\docType{class}
\name{stringCoverageGenotypeList-class}
\alias{stringCoverageGenotypeList-class}
\title{Combined stringCoverage- and genotypeIdentifiedList}
\description{
Merges a stringCoverageList with a genotypeIdentifiedList.
}
