% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{stringCoverageGenotypeList}
\alias{stringCoverageGenotypeList}
\title{Combined string coverage and genotype information}
\format{
A list of \link[tibble]{tibble}'s one for each of the 10 markers containing the combined string coverage and genotypic information.
}
\usage{
data("stringCoverageGenotypeList")
}
\description{
A merge of the \link{stringCoverageList} and \link{genotypeList} data.
}
\author{
Søren B. Vilsen \email{svilsen@math.aau.dk}
}
\keyword{data}
