% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\docType{class}
\name{stringCoverageList-class}
\alias{stringCoverageList-class}
\title{A string coverage list}
\description{
A list of tibbles, one for every marker, used to contain the sequencing information of STR MPS data.
The tibbles should include columns with the following names: "Marker", "BasePairs", "Allele", "Type", "MotifLength", "ForwardFlank", "Region", "ReverseFlank", "Coverage", "AggregateQuality", and "Quality".
}
