% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\docType{data}
\name{stringCoverageList}
\alias{stringCoverageList}
\title{Aggregated string coverage.}
\format{
A list of \link[tibble]{tibble}'s one for each of the 10 markers, showing the aggregated information on a string-by-string basis.
}
\usage{
data("stringCoverageList")
}
\description{
The aggregated string coverage of the \link{identifiedSTRs} data, created by the  \link{stringCoverage} function.
}
\author{
Søren B. Vilsen \email{svilsen@math.aau.dk}
}
\keyword{data}
