% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\docType{class}
\name{stringCoverageNoiseList-class}
\alias{stringCoverageNoiseList-class}
\title{Combined stringCoverage- and noiseIdentifiedList}
\description{
Merges a stringCoverageList with a noiseIdentifiedList
}
