% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_workflow_fun.R
\name{STRMPSWorkflowBatch}
\alias{STRMPSWorkflowBatch}
\title{Batch wrapper for the workflow function}
\usage{
STRMPSWorkflowBatch(
  input,
  output,
  ignorePattern = NULL,
  continueCheckpoint = NULL,
  control = workflow.control()
)
}
\arguments{
\item{input}{A directory where fastq input-files are stored.}

\item{output}{A directory where output-files are stored.}

\item{ignorePattern}{A pattern parsed to grepl used to filter input strings.}

\item{continueCheckpoint}{Choose a checkpoint to continue from in the workflow. If NULL the function will run the entire workflow.}

\item{control}{Function controlling non-crucial parameters and other control functions.}
}
\value{
If 'output' not provided the function simply returns the stringCoverageList-object.
If an output is provided the function will store ALL created objects at the output-path, i.e. nothing is returned.
}
\description{
The function takes an input directory and performs the entire analysis workflow described in (ADD REF).
The function creates a series of objects needed for further analyses and stores them at the output location.
}
