% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_workflow_fun.R
\name{STRMPSWorkflowCollectStutters}
\alias{STRMPSWorkflowCollectStutters}
\title{Collect stutters files}
\usage{
STRMPSWorkflowCollectStutters(stutterDirectory, storeCollection = TRUE)
}
\arguments{
\item{stutterDirectory}{The out most directory containing all stutter files to be collected.}

\item{storeCollection}{TRUE/FALSE: Should the collected tibble be stored? If 'FALSE' the tibble is returned.}
}
\value{
If 'storeCollection' is TRUE nothing is returned, else the stutter collection is returned.
}
\description{
Collects all stutter files created by the batch version of the \link{STRMPSWorkflow} function.
}
