% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygote_balance.R
\name{trueAlleles}
\alias{trueAlleles}
\title{Creates trueAllelesList-object.}
\usage{
trueAlleles(alleles, strings, markers, motifLength)
}
\arguments{
\item{alleles}{the true alleles provided as a list of numerics or a vector of strings with the alleles seperated by a ','.}

\item{strings}{the true strings of an individual (has to have same ordering as alleles) provided as 'character'.}

\item{markers}{the markers needed (use e.g. names of stringCoverage-object).}

\item{motifLength}{the length of the motif.}
}
\value{
An object of class \code{trueAllelesList}.
}
\description{
The function creates a list, with an element for each marker, containing two lists, with the true alleles and strings, respectively.
}

