% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vmatchMultiPattern}
\alias{vmatchMultiPattern}
\title{vmatchMultiPattern}
\usage{
vmatchMultiPattern(pattern, subject, max_mismatch, min_mismatch, with_indels,
  fixed, algorithm, matches_as, envir)
}
\arguments{
\item{pattern}{an XStringSet-object.}

\item{subject}{an XStringSet-object.}

\item{max_mismatch}{the maximum allowed number of mismatching bases between the pattern and the subject.}

\item{min_mismatch}{the minimum allowed number of mismatching bases between the pattern and the subject.}
}
\description{
Wrapper function for \code{XStringSet_vmatch_pattern} allowing for multiple patterns.
Alternate version of vmatchMultiPattern.
}
\details{
\code{vmatchMultiPattern}...
}

