% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_workflow_fun.R
\name{workflow}
\alias{workflow}
\title{Workflow function}
\usage{
workflow(input, output = NULL, continueCheckpoint = NULL,
  flankingRegions = NULL, control = workflow.control())
}
\arguments{
\item{input}{A path to a \code{.fastq}-file.}

\item{output}{A directory where output-files are stored.}

\item{continueCheckpoint}{Choose a checkpoint to continue from in the workflow. If NULL the function will run the entire workflow.}

\item{flankingRegions}{The flanking regions used to identify markers and trim alleles.}

\item{control}{Function controlling non-crucial parameters and other control functions.}
}
\value{
If 'output' not provided the function simply returns the stringCoverageList-object.
If an output is provided the function will store ALL created objects at the output-path, but nothing is returned.
}
\description{
The function takes an input file and performs the entire analysis workflow described in (ADD REF).
The function creates a series of objects needed for further analyses.
An output folder can be provided to store the objects as \code{.RData}-files.
}
