% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_workflow_fun.R
\name{workflowList}
\alias{workflowList}
\title{Workflow function for multiple files.}
\usage{
workflowList(input, output = NULL, flankingRegions = NULL,
  control = workflow.control())
}
\arguments{
\item{input}{A vector of paths to \code{.fastq}-files.}

\item{output}{A directory where output-files are stored (forced).}

\item{flankingRegions}{The flanking regions used to identify markers and trim alleles.}

\item{control}{Function controlling non-crucial parameters and other control functions.}
}
\value{
The function will store ALL created objects at the output-path; nothing is returned.
}
\description{
The function wrap-function for the \link{workflow}-function allowing for more than one input file.
}
