% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stutter_functions.R
\name{findLeftShoulder}
\alias{findLeftShoulder}
\title{Find left shoulder}
\usage{
findLeftShoulder(stringCoverageGenotypeListObject, motifLength, searchDirection,
  gapOpeningPenalty = 6, gapExtensionPenalty = 1)
}

