% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STR_Region_Identification.R
\name{identifySTRRegions.control}
\alias{identifySTRRegions.control}
\title{Control function for identifySTRRegions}
\usage{
identifySTRRegions.control(colList = NULL, numberOfThreads = 4L,
  reversed = TRUE, includeReverseComplement = FALSE, combineLists = TRUE,
  removeEmptyMarkers = TRUE, matchPatternMethod = "mclapply")
}
\value{
A list containing...
}
\description{
\code{identifySTRRegions.control}
}
