% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STR_Region_Identification.R
\name{identifySTRRegions.control}
\alias{identifySTRRegions.control}
\title{Control function for identifySTRRegions}
\usage{
identifySTRRegions.control(colList = NULL, numberOfThreads = 4L,
  reversed = TRUE, includeReverseComplement = TRUE, combineLists = TRUE,
  removeEmptyMarkers = TRUE, matchPatternMethod = "vmatch")
}
\value{
A list containing...
}
\description{
\code{identifySTRRegions.control}
}

