% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinb_noise_threshold.R
\name{noiseThresholds}
\alias{noiseThresholds}
\title{Sets Noise Threshold}
\usage{
noiseThresholds(stringCoverageListObject, predefinedProportions,
  numberOfStandardDeviations, numberOfThreads)
}
\arguments{
\item{numberOfThreads}{}
}
\value{
A set of thresholds for noise identification.
}
\description{
The function creates noise thresholds based on the provided stringCoverageList-object, and a set of pre-defined proportions, using either a \link{kinb} or a geometric distribution.
}
