% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_genotype_functions.R
\name{stringCoverage.control}
\alias{stringCoverage.control}
\title{String coverage coontrol object}
\usage{
stringCoverage.control(
  numberOfThreads = 4L,
  simpleReturn = TRUE,
  uniquelyAssigned = TRUE,
  additionalFlags = FALSE,
  numberOfMutations = 1,
  includeAverageBaseQuality = FALSE,
  meanFunction = mean,
  trace = FALSE
)
}
\arguments{
\item{numberOfThreads}{The number of cores used for parallelisation.}

\item{simpleReturn}{TRUE/FALSE: Should the returned object be simplified?}

\item{uniquelyAssigned}{TRUE/FALSE: Should regions which are not uniquely assigned be removed?}

\item{additionalFlags}{TRUE/FALSE: Create additional flags for assessing sequence reliability?}

\item{numberOfMutations}{the maximum number of mutations (base-calling errors) allowed during flanking region identification.}

\item{includeAverageBaseQuality}{Should the average base quality of the region be included?}

\item{meanFunction}{The function used to average the base qualities.}

\item{trace}{TRUE/FALSE: Show trace?}
}
\value{
List of parameters used for the 'stringCoverage' function.
}
\description{
String coverage coontrol object
}
\details{
Control function for the 'stringCoverage' function. Sets default values for the parameters.
}
