% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vmatchMultiPatternSeqAn}
\alias{vmatchMultiPatternSeqAn}
\title{vmatchMultiPatternSeqAn}
\usage{
vmatchMultiPatternSeqAn(pattern, subject, max_mismatch)
}
\arguments{
\item{pattern}{a vector of strings.}

\item{subject}{a vector of strings.}

\item{max_mismatch}{the maximum allowed number of mismatching bases between the pattern and the subject.}
}
\description{
Function achieving the same as the wrapper functions for \code{XStringSet_vmatch_pattern} allowing for multiple patterns.
This alternate method is based on the external SeqAn C++ library.
}
\details{
\code{vmatchMultiPatternSeqAn}.
}

